Support pre-sm_60 GPUs (GTX 780) with fixed-point energy accumulation
CUDA architecture targets expanded to sm_35, sm_52, sm_89 to support
GTX 780 (Kepler), GTX 1060/1070 (Maxwell/Pascal), and RTX 4090 (Ada).

Isolated force kernels (Bonded, Nonbonded, Site) and HMC/NUTS integrators
switched from float atomicAdd (requires sm_60+) to fixed-point unsigned
long long accumulators for energy reduction, matching OpenMM's own
approach for backward compatibility.

CUDA kernel changes:
- nullptr comparisons replaced with 0 (device code compatibility)
- Device lambdas replaced with __device__ helper functions
- GBSA grid generation: extracted computeHCTSumAtPoint as standalone
  __device__ function

+++ platforms/cuda/src/CudaMultiGroupHMCKernels.h
@@ -75,7 +75,7 @@ private:
     OpenMM::CudaArray acceptedBuffer;
 
     // Host-side caches
-    std::vector<double> groupKEHost;
+    std::vector<unsigned long long> groupKEHost;
     std::vector<double> groupStepSizesHost;
     std::vector<double> groupKTHost;
     std::vector<int> acceptedHost;
@@ -108,14 +108,14 @@ private:
     std::normal_distribution<double> normalDist;
 
     // MC GPU buffers
-    OpenMM::CudaArray groupCOMBuffer;       // 4 doubles per group (sum_mx, sum_my, sum_mz, sum_m)
+    OpenMM::CudaArray groupCOMBuffer;       // 4 unsigned long longs per group (fixed-point sum_mx, sum_my, sum_mz, sum_m)
     OpenMM::CudaArray mcEnabledBuffer;      // int per group
     OpenMM::CudaArray mcRotationBuffer;     // 9 doubles per group (3x3 row-major)
     OpenMM::CudaArray mcTranslationBuffer;  // 3 doubles per group
     OpenMM::CudaArray mcCOMBuffer;          // 3 doubles per group (finalized COM)
 
     // MC host-side vectors
-    std::vector<double> groupCOMHost;        // 4 * numGroups
+    std::vector<unsigned long long> groupCOMHost; // 4 * numGroups (fixed-point)
     std::vector<int> mcEnabledHost;
     std::vector<double> mcRotationHost;      // 9 * numGroups
     std::vector<double> mcTranslationHost;   // 3 * numGroups
